Rewriting RequestMethod attributes.
Doing it that way, i can use their common base,
RequestMethod, as return type for modules.
This allows uniform method type.
Adding a readme.txt there as well.

I will also move the constants for request methods there,
their place is not in the request itself, as they
are already defined when the module accepts it.

There will be routers, which will take modules for given url-s.
They will load the modules on demand, and will run the startups in order.
A startup will not be called if it's module is not loaded.
A submodule path that collides with a module path will be ignored.